+++ src/layouts/AppLayout.tsx
@@ -1,4 +1,4 @@
-import Header from "./Header";
+import Header from "../components/layout/Header";
 import { Outlet } from "react-router-dom";
 
 export default function AppLayout() {
